use a module to avoid duplicated code
